refactor: Split Tier 1 large source files into focused modules
Split the 5 largest source files (>800 lines each) into smaller,
focused modules for better readability and maintainability.

- AppShell.tsx (1,871→360 lines): Extract 10 modules —
  useWorkspaceLoader, useWorkspaceRestore, useTabOperations,
  useTabPersistence, useTileDragDrop, TileRenderer,
  SidebarElements, AppShellDialogs, DesktopLayout, MobileLayout
- plugins.ts (1,118 lines): Split into keyboardPlugins.ts and
  inputPlugins.ts, delete original
- AgentEditorPanel.tsx (833→290 lines): Extract AgentInfoCard,
  EditorSettingsDropdown, controlResponseHandling
- MarkdownEditor.tsx (932→500 lines): Extract editorSetup,
  draftManagement, editorRefHandlers
- messageRenderers.tsx (833→520 lines): Extract planModeRenderers,
  taskRenderers, rendererUtils
- toolRenderers.tsx (602→470 lines): Extract toolDetailRenderers
- Fix pre-existing lint warnings in TabBar.test.tsx and
  WorkspaceContextMenu.test.tsx
